remove existing response paged only for project and make consistent paging response
