+++ apps/api/src/main/kotlin/org/aiml/api/dto/project/ProjectResponse.kt
@@ -41,33 +41,3 @@ data class ProjectUserResponse(
     )
   }
 }
-
-data class ProjectPageResponse(
-  val projects: List<ProjectBaseResponse>,
-  val pageInfo: PageInfo
-) {
-  companion object {
-    fun from(page: Page<ProjectDTO>): ProjectPageResponse {
-      val projects = page.content.map { ProjectBaseResponse.from(it) }
-      val pageInfo = PageInfo(
-        page = page.number,
-        size = page.size,
-        hasNext = page.hasNext(),
-        hasPrevious = page.hasPrevious(),
-        sort = page.sort.toString()
-      )
-      return ProjectPageResponse(
-          projects = projects,
-          pageInfo = pageInfo
-      )
-    }
-  }
-}
-
-data class PageInfo (
-  val page: Int,                // 현재 페이지 번호 (0부터 시작)
-  val size: Int,                // 페이지 요청 시 크기
-  val hasNext: Boolean,         // 다음 페이지 존재 여부
-  val hasPrevious: Boolean,     // 이전 페이지 존재 여부
-  val sort: String              // 정렬 기준
-)
